requirements.txt
azure-identity==1.14
msgraph-sdk==1.0.0a16
pandas==2.1.4
dotenv==1.0.0